pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "subtitle-studio"
version = "0.1.0"
description = "Unified CLI pipeline for YouTube subtitles — extraction, verification and translation."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Taverne Tech", email = "tavernetech@gmail.com" },
]
keywords = [
    "subtitles",
    "youtube",
    "whisper",
    "transcription",
    "translation",
    "srt",
    "ffmpeg",
    "anthropic",
    "claude",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    # generation
    "faster-whisper>=1.0",
    "stable-ts>=2.19",
    "srt>=3.5",
    # detection
    "anthropic>=0.86.0",
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
api = ["openai>=1.0"]  # backend Whisper cloud

[project.urls]
Homepage = "https://taverne.tech"
Repository = "https://github.com/DrakkarStorm/subtitle-studio"
Issues = "https://github.com/DrakkarStorm/subtitle-studio/issues"

[project.scripts]
subtitle-studio = "subtitle_studio.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["subtitle_studio"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "pre-commit>=3.7",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = ["-ra", "--strict-markers", "--strict-config"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
# Typer's @app.command() decorator is untyped upstream — safe to ignore here.
disable_error_code = ["untyped-decorator"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "B008",   # typer.Argument/Option en default — pattern officiel Typer
]
